refactor(ux-tests): share fixture loading helpers (#0000)
