+++ crates/perl-lsp-ux-tests/tests/ux_scenario_28_mojolicious_completion_ranking.rs
@@ -10,24 +10,17 @@
 //! - generated-candidate and provenance-label coverage
 //! - dynamic/fallback label coverage when the provider exposes it
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 use perl_lsp_ux_tests::{
-    ScenarioConfig, UxCiTier, UxComponent, UxHarness, UxScenarioSkip, run_ux_scenario,
+    FixtureFile, ScenarioConfig, UxCiTier, UxComponent, UxHarness, binary_available,
+    load_real_project_fixture_files, missing_binary_skip, run_ux_scenario,
 };
 use serde::Serialize;
 use serde_json::Value;
-use std::fs;
-use std::path::{Path, PathBuf};
 use std::time::Duration;
 
 const SCENARIO_FILE: &str = "ux_scenario_28_mojolicious_completion_ranking.rs";
 const TOP_N: usize = 10;
-
-#[derive(Debug)]
-struct FixtureFile {
-    relative_path: String,
-    content: String,
-}
 
 #[derive(Debug)]
 struct CompletionProbe {
@@ -56,60 +49,6 @@ struct CompletionProbeReport {
     generated_candidate_hits: Vec<String>,
     generated_provenance_label_hits: Vec<String>,
     dynamic_or_fallback_label_hits: Vec<String>,
-}
-
-fn binary_available() -> bool {
-    perl_lsp_ux_tests::resolve_binary().is_ok()
-}
-
-fn missing_binary_skip() -> UxScenarioSkip {
-    UxScenarioSkip::infra("PERL_LSP_BIN not set and target/debug/perl-lsp not found")
-}
-
-fn workspace_root() -> Result<PathBuf> {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(Path::parent)
-        .map(Path::to_path_buf)
-        .context("CARGO_MANIFEST_DIR must be nested under the workspace root")
-}
-
-fn mojolicious_fixture_root() -> Result<PathBuf> {
-    Ok(workspace_root()?.join("test_corpus").join("real_projects").join("mojolicious_skeleton"))
-}
-
-fn is_perl_source(path: &Path) -> bool {
-    path.extension()
-        .and_then(|extension| extension.to_str())
-        .is_some_and(|extension| matches!(extension, "pm" | "pl" | "t"))
-}
-
-fn collect_perl_files(root: &Path, dir: &Path, files: &mut Vec<FixtureFile>) -> Result<()> {
-    for entry in fs::read_dir(dir).with_context(|| format!("reading {}", dir.display()))? {
-        let entry = entry.with_context(|| format!("reading an entry under {}", dir.display()))?;
-        let path = entry.path();
-        if path.is_dir() {
-            collect_perl_files(root, &path, files)?;
-        } else if is_perl_source(&path) {
-            let relative_path = path
-                .strip_prefix(root)
-                .with_context(|| format!("stripping fixture root from {}", path.display()))?
-                .to_string_lossy()
-                .replace('\\', "/");
-            let content =
-                fs::read_to_string(&path).with_context(|| format!("reading {}", path.display()))?;
-            files.push(FixtureFile { relative_path, content });
-        }
-    }
-    Ok(())
-}
-
-fn load_mojolicious_fixture_files() -> Result<Vec<FixtureFile>> {
-    let root = mojolicious_fixture_root()?;
-    let mut files = Vec::new();
-    collect_perl_files(&root, &root, &mut files)?;
-    files.sort_by(|left, right| left.relative_path.cmp(&right.relative_path));
-    Ok(files)
 }
 
 fn create_mojolicious_harness(files: &[FixtureFile]) -> Result<UxHarness> {
@@ -306,7 +245,7 @@ fn scenario_28_mojolicious_visible_symbol_ranking_receipt() {
                 return Err(missing_binary_skip().into());
             }
 
-            let fixture_files = load_mojolicious_fixture_files()?;
+            let fixture_files = load_real_project_fixture_files("mojolicious_skeleton")?;
             recorder
                 .check("mojolicious fixture has committed Perl files", !fixture_files.is_empty())?;
